Refactor: detailmodal 컴포넌트 분리
